Added delay when leaving a room

+++ Assets/Scripts/Componets/WaitComponent.cs
@@ -3,6 +3,6 @@
 public struct WaitComponent : IComponentData
 {
     // time is in unix format and needs to be that way to make it DOTS compliant, since the class Time is not suitable as a component field
-    public long waitEndTime;
+    public float waitEndTime;
     public int slotsToWait;
 }
